requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
slack-sdk==3.27.1
slack-bolt==1.18.1
schedule==1.2.1
requests==2.31.0
openai==1.12.0
ffmpeg-python==0.2.0
whisper==1.1.10
gunicorn==21.2.0
